Move all trajectory calculators to new parent class.
