+++ mdsuite/calculators/einstein_helfand_thermal_kinaci.py
@@ -26,23 +26,30 @@ Summary
 MDSuite module for the computation of the thermal conductivity in solids using the
 Einstein method as applied to the Kinaci integrated thermal flux.
 """
+from abc import ABC
+
 import numpy as np
 import tensorflow as tf
 from tqdm import tqdm
-from mdsuite.calculators.calculator import Calculator, call
+from mdsuite.calculators.calculator import call
 from dataclasses import dataclass
 from mdsuite.database import simulation_properties
+from mdsuite.calculators import TrajectoryCalculator
+from mdsuite.utils.calculator_helper_methods import fit_einstein_curve
 
 
 @dataclass
 class Args:
+    """
+    Data class for the saved properties.
+    """
     data_range: int
     correlation_time: int
     tau_values: np.s_
     atom_selection: np.s_
 
 
-class EinsteinHelfandThermalKinaci(Calculator):
+class EinsteinHelfandThermalKinaci(TrajectoryCalculator, ABC):
     """
     Class for the Einstein-Helfand Ionic Conductivity
 
@@ -93,16 +100,15 @@ class EinsteinHelfandThermalKinaci(Calculator):
         self.y_label = r"$$\text{MSD} / m^2/s$$"
         self.analysis_name = "Einstein Helfand Thermal Conductivity Kinaci"
 
-        self.database_group = "Thermal_Conductivity"
+        self._dtype = tf.float64
 
-        self.prefactor: float
+        self.prefactor = None
 
     @call
     def __call__(
             self,
             plot=True,
             data_range=500,
-            save=True,
             correlation_time=1,
             tau_values: np.s_ = np.s_[:],
             gpu: bool = False,
@@ -116,8 +122,7 @@ class EinsteinHelfandThermalKinaci(Calculator):
                 if true, plot the output.
         data_range : int
                 Data range to use in the analysis.
-        save : bool
-                if true, save the output.
+
         correlation_time : int
                 Correlation time to use in the window sampling.
         gpu : bool
@@ -132,17 +137,24 @@ class EinsteinHelfandThermalKinaci(Calculator):
             atom_selection=np.s_[:],
         )
 
+        self.plot = plot
         self.gpu = gpu
         self.time = self._handle_tau_values()
         self.msd_array = np.zeros(self.data_resolution)
 
-    def _calculate_prefactor(self, species: str = None):
+    def check_input(self):
+        """
+        Check the user input to ensure no conflicts are present.
+
+        Returns
+        -------
+
+        """
+        self._run_dependency_check()
+
+    def _calculate_prefactor(self):
         """
         Compute the ionic conductivity prefactor.
-
-        Parameters
-        ----------
-        species
 
         Returns
         -------
@@ -154,7 +166,7 @@ class EinsteinHelfandThermalKinaci(Calculator):
             6
             * self.experiment.volume
             * self.experiment.temperature
-            * self.experiment.units["boltzman"]
+            * self.experiment.units["boltzmann"]
         )
         units_change = (
             self.experiment.units["energy"]
@@ -190,14 +202,14 @@ class EinsteinHelfandThermalKinaci(Calculator):
         msd = self.prefactor * tf.reduce_sum(msd, axis=1)
         self.msd_array += np.array(msd)  # Update the averaged function
 
-    def _post_operation_processes(self, species: str = None):
+    def _post_operation_processes(self):
         """
         call the post-op processes
         Returns
         -------
 
         """
-        result = self._fit_einstein_curve([self.time, self.msd_array])
+        result = fit_einstein_curve([self.time, self.msd_array])
 
         data = {
             "thermal_conductivity": result[0],
@@ -224,6 +236,7 @@ class EinsteinHelfandThermalKinaci(Calculator):
         -------
 
         """
+        self.check_input()
         # Compute the pre-factor early.
         self._calculate_prefactor()
