+++ mdsuite/calculators/green_kubo_self_diffusion_coefficients.py
@@ -23,22 +23,28 @@ If you use this module please cite us with:
 
 Summary
 -------
+Module for the computation of diffusion coefficients using the Green-Kubo approach.
 """
+from abc import ABC
+
 import numpy as np
 import tensorflow_probability as tfp
 import tensorflow as tf
 from bokeh.models import Span
 from tqdm import tqdm
 from typing import Union
-from mdsuite.calculators.calculator import Calculator, call
-from mdsuite.database.scheme import Computation
+from mdsuite.calculators.calculator import call
 from dataclasses import dataclass
 from mdsuite.database import simulation_properties
 from typing import List, Any
+from mdsuite.calculators import TrajectoryCalculator
 
 
 @dataclass
 class Args:
+    """
+    Data class for the saved properties.
+    """
     data_range: int
     correlation_time: int
     atom_selection: np.s_
@@ -48,7 +54,7 @@ class Args:
     integration_range: int
 
 
-class GreenKuboDiffusionCoefficients(Calculator):
+class GreenKuboDiffusionCoefficients(TrajectoryCalculator, ABC):
     """
     Class for the Green-Kubo diffusion coefficient implementation
     Attributes
@@ -91,26 +97,26 @@ class GreenKuboDiffusionCoefficients(Calculator):
         self.loaded_property = simulation_properties.velocities
         self.scale_function = {"linear": {"scale_factor": 150}}
 
-        self.database_group = "Diffusion_Coefficients"
         self.x_label = r"$$\text{Time} / s$$"
         self.y_label = r"$$\text{VACF} / m^{2}/s^{2}$$"
         self.analysis_name = "Green Kubo Self-Diffusion Coefficients"
         self.result_keys = ["diffusion_coefficient", "uncertainty"]
         self.result_series_keys = ["time", "acf"]
 
+        self._dtype = tf.float64
+
     @call
     def __call__(
-        self,
-        plot: bool = True,
-        species: list = None,
-        data_range: int = 500,
-        save: bool = True,
-        correlation_time: int = 1,
-        atom_selection=np.s_[:],
-        molecules: bool = False,
-        gpu: bool = False,
-        tau_values: Union[int, List, Any] = np.s_[:],
-        integration_range: int = None,
+            self,
+            plot: bool = True,
+            species: list = None,
+            data_range: int = 500,
+            correlation_time: int = 1,
+            atom_selection=np.s_[:],
+            molecules: bool = False,
+            gpu: bool = False,
+            tau_values: Union[int, List, Any] = np.s_[:],
+            integration_range: int = None,
     ):
         """
         Constructor for the Green-Kubo diffusion coefficients class.
@@ -123,8 +129,6 @@ class GreenKuboDiffusionCoefficients(Calculator):
                 List of species on which to operate.
         data_range : int
                 Data range to use in the analysis.
-        save : bool
-                if true, save the output.
         correlation_time : int
                 Correlation time to use in the window sampling.
         atom_selection : np.s_
@@ -160,8 +164,18 @@ class GreenKuboDiffusionCoefficients(Calculator):
         self.gpu = gpu
         self.plot = plot
         self.time = self._handle_tau_values()
-        self.vacf = np.zeros(self.data_resolution)
+        self.vacf = tf.zeros(self.data_resolution)
         self.sigma = []
+
+    def check_input(self):
+        """
+        Check the user input to ensure no conflicts are present.
+
+        Returns
+        -------
+
+        """
+        self._run_dependency_check()
 
     def _calculate_prefactor(self, species: str = None):
         """
@@ -180,19 +194,19 @@ class GreenKuboDiffusionCoefficients(Calculator):
         if self.args.molecules:
             numerator = self.experiment.units["length"] ** 2
             denominator = (
-                3
-                * self.experiment.units["time"]
-                * (self.args.integration_range - 1)
-                * len(self.experiment.molecules[species]["indices"])
+                    3
+                    * self.experiment.units["time"]
+                    * (self.args.integration_range - 1)
+                    * len(self.experiment.molecules[species]["indices"])
             )
             self.prefactor = numerator / denominator
         else:
             numerator = self.experiment.units["length"] ** 2
             denominator = (
-                3
-                * self.experiment.units["time"]
-                * self.args.integration_range
-                * len(self.experiment.species[species]["indices"])
+                    3
+                    * self.experiment.units["time"]
+                    * self.args.integration_range
+                    * len(self.experiment.species[species]["indices"])
             )
             self.prefactor = numerator / denominator
 
@@ -216,24 +230,32 @@ class GreenKuboDiffusionCoefficients(Calculator):
         self.vacf += vacf
         self.sigma.append(
             np.trapz(
-                vacf[: self.args.integration_range], x=self.time[: self.args.integration_range]
+                vacf[: self.args.integration_range],
+                x=self.time[: self.args.integration_range]
             )
         )
 
-    def plot_data(self, data):
-        """Plot the data"""
+    def plot_data(self, data: dict):
+        """
+        Plot the data
+
+        Parameters
+        ----------
+        data : dict
+                Data loaded from the sql database to be plotted.
+        """
         for selected_species, val in data.items():
             span = Span(
                 location=(
-                    np.array(val[self.result_series_keys[0]])
-                    * self.experiment.units["time"]
+                        np.array(val[self.result_series_keys[0]])
+                        * self.experiment.units["time"]
                 )[self.args.integration_range - 1],
                 dimension="height",
                 line_dash="dashed",
             )
             self.run_visualization(
                 x_data=np.array(val[self.result_series_keys[0]])
-                * self.experiment.units["time"],
+                       * self.experiment.units["time"],
                 y_data=np.array(val[self.result_series_keys[1]]),
                 title=(
                     f"{val[self.result_keys[0]]: 0.3E} +-"
@@ -274,6 +296,7 @@ class GreenKuboDiffusionCoefficients(Calculator):
         -------
 
         """
+        self.check_input()
         # Loop over species
         for species in self.args.species:
             dict_ref = str.encode("/".join([species, self.loaded_property[0]]))
@@ -282,11 +305,11 @@ class GreenKuboDiffusionCoefficients(Calculator):
             batch_ds = self.get_batch_dataset([species])
 
             for batch in tqdm(
-                batch_ds,
-                ncols=70,
-                desc=species,
-                total=self.n_batches,
-                disable=self.memory_manager.minibatch,
+                    batch_ds,
+                    ncols=70,
+                    desc=species,
+                    total=self.n_batches,
+                    disable=self.memory_manager.minibatch,
             ):
                 ensemble_ds = self.get_ensemble_dataset(batch, species, split=True)
