+++ mdsuite/memory_management/__init__.py
@@ -24,3 +24,6 @@ If you use this module please cite us with:
 Summary
 -------
 """
+from .memory_manager import MemoryManager
+
+__all__ = ["MemoryManager"]
